# apps/command/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Team VIT, SIH 2026 PS SIH26085"
    },
    "description": "Street-level urban flood nowcasting digital twin (SIH 2026, PS SIH26085, MoES). Every response carries `run_id` and `valid_ts` where a run is involved; errors use `{ error: { code, message, run_id } }`. Endpoints answering 501 are contract-only until their engine's phase lands.",
    "title": "VARUNA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/feeds/road-conditions": {
      "get": {
        "description": "The provider feed: every segment this run predicts will stop the given vehicle.\n\nOne feature per segment, with the window it is impassable for. A navigation app does not want\na depth in centimetres, it wants \"closed from 08:20 to 10:05\", so that is what this carries -\nwith the depth beside it for anyone who does.",
        "operationId": "road_conditions_v1_feeds_road_conditions_get",
        "parameters": [
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "default": "car",
              "title": "Profile",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "default": "mumbai",
              "title": "City",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Road Conditions V1 Feeds Road Conditions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Impassable and degraded roads, as GeoJSON",
        "tags": [
          "route"
        ]
      }
    },
...
    "/v1/reachability": {
      "get": {
        "description": "Isochrones and the collapse flag.",
        "operationId": "reachability_v1_reachability_get",
        "parameters": [
          {
            "description": "asset_id or exact name from /v1/route/facilities",
            "in": "query",
            "name": "facility",
            "required": true,
            "schema": {
              "description": "asset_id or exact name from /v1/route/facilities",
              "title": "Facility",
              "type": "string"
            }
          },
          {
            "description": "Instant to measure at, ISO 8601.",
            "in": "query",
            "name": "t",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instant to measure at, ISO 8601.",
              "title": "T"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "default": "ambulance",
              "title": "Profile",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "default": "mumbai",
              "title": "City",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reachability V1 Reachability Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One facility's 5/10/15-minute catchment, against dry",
        "tags": [
          "route"
        ]
      }
    },
...
    "/v1/route": {
      "post": {
        "description": "Plan one trip.\n\nBody: ``{origin, destination, depart_at?, profile?, risk_tolerance?, run_id?, spread?,\ntrip_id?, explain?}``.\n\nReturns both routes, because the comparison is the product: a dispatcher who only sees the\nsafe route has no way to judge whether the detour was worth it.\n\n``spread`` (default true) adds ``corridors[]`` - up to three safe roads with the share of\ntraffic the policy gives each, and the one this request is assigned to. ``trip_id`` makes\nthat assignment stable for a trip, so a reader who reloads is not sent somewhere else; it is\nthe client's own random id and nothing is stored against it. ``explain`` (default true) adds\n``reasons[]``, which is structured data - the frontend writes the sentences (TECH_SPEC 3.2).",
        "operationId": "route_v1_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Route V1 Route Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route around the forecast water, beside what a naive router does",
        "tags": [
          "route"
        ]
      }
    },
    "/v1/route/facilities": {
      "get": {
        "description": "The pickable facilities, from the city's own asset layer.",
        "operationId": "route_facilities_v1_route_facilities_get",
        "parameters": [
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "default": "mumbai",
              "title": "City",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Route Facilities V1 Route Facilities Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hospitals and fire stations the isochrones can start at",
        "tags": [
          "route"
        ]
      }
    },
...
  },
...
}